refactor: Created a new subdivision folder for database manipulation. Renamed database.js to dbInit.js.
