+++ index.html
@@ -13,9 +13,11 @@
     <script type="module" src="js/utils/dateUtils.js"></script> 
     <script type="module" src="js/utils/colorUtils.js"></script> 
     <script type="module" src="js/utils/validationUtils.js"></script> 
-    <script type="module" src="js/database.js"></script> 
-    <script type="module" src="js/Game.js"></script> 
-    <script type="module" src="js/Task.js"></script> 
+    <script type="module" src="js/database/dbInit.js"></script> 
+    <script type="module" src="js/database/gameDB.js"></script> 
+    <script type="module" src="js/database/taskDB.js"></script> 
+    <script type="module" src="js/data/Game.js"></script> 
+    <script type="module" src="js/data/Task.js"></script> 
     <script type="module" src="js/index.js"></script> 
     <title>Gacha Management</title>
 </head>
